All fields, variables, parameters, and return values that hold a character's total subpoints of any stat have been changed from int to long.  This should allow most KoLmafia features to work with characters up to level 255 for now, and much higher than that if the server-side level cap is ever raised.  Known limitations are:
* Substats until the next level are still represented as ints in several
places.  That won't overflow until somewhat beyond level 800.

* Level and stat conditions in the adventure goals field won't work beyond
level 216.

* The ASH my_basestat() function, when passed one of the constants that
request substat values, still return 32-bit signed values - which will go
negative past level 216.  However, the difference between two substat
readings will still accurately indicate the number of substats gained between
those readings, whether or not the wrap-around occurred.  (This is not
necessarily the final disposition of this function.)


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7299 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/KoLmafiaCLI.java
@@ -5141,8 +5141,8 @@ public class KoLmafiaCLI
 
 			int primeIndex = KoLCharacter.getPrimeIndex();
 
-			AdventureResult.CONDITION_SUBSTATS[ primeIndex ] =
-				KoLCharacter.calculateSubpoints( ( level - 1 ) * ( level - 1 ) + 4, 0 ) - KoLCharacter.getTotalPrime();
+			AdventureResult.CONDITION_SUBSTATS[ primeIndex ] = (int)
+				(KoLCharacter.calculateSubpoints( ( level - 1 ) * ( level - 1 ) + 4, 0 ) - KoLCharacter.getTotalPrime());
 
 			return AdventureResult.CONDITION_SUBSTATS_RESULT;
 		}
@@ -5154,11 +5154,11 @@ public class KoLmafiaCLI
 			int points = StringUtilities.parseInt( splitCondition[ 0 ] );
 			int statIndex = conditionString.indexOf( "mus" ) != -1 ? 0 : conditionString.indexOf( "mys" ) != -1 ? 1 : 2;
 
-			AdventureResult.CONDITION_SUBSTATS[ statIndex ] = KoLCharacter.calculateSubpoints( points, 0 );
+			AdventureResult.CONDITION_SUBSTATS[ statIndex ] = (int) KoLCharacter.calculateSubpoints( points, 0 );
 			AdventureResult.CONDITION_SUBSTATS[ statIndex ] =
 				Math.max(
 					0,
-					AdventureResult.CONDITION_SUBSTATS[ statIndex ] - ( conditionString.indexOf( "mus" ) != -1 ? KoLCharacter.getTotalMuscle() : conditionString.indexOf( "mys" ) != -1 ? KoLCharacter.getTotalMysticality() : KoLCharacter.getTotalMoxie() ) );
+					AdventureResult.CONDITION_SUBSTATS[ statIndex ] - (int) ( conditionString.indexOf( "mus" ) != -1 ? KoLCharacter.getTotalMuscle() : conditionString.indexOf( "mys" ) != -1 ? KoLCharacter.getTotalMysticality() : KoLCharacter.getTotalMoxie() ) );
 
 			return AdventureResult.CONDITION_SUBSTATS_RESULT;
 		}
